+++ src/main/kotlin/com/admin_bot/features/authentification/model/RefreshTokenStorage.kt
@@ -4,7 +4,7 @@ interface RefreshTokenStorage {
     /**
      * Returns true if [refreshToken] was in storage and was deleted
      */
-     suspend fun deleteRefreshTokenIfContains(refreshToken: String, botId: Int): Boolean
+     suspend fun deleteRefreshTokenIfContains(refreshToken: String, botId: Long): Boolean
 
-     suspend fun saveRefreshToken(refreshToken: String, botId: Int)
+     suspend fun saveRefreshToken(refreshToken: String, botId: Long)
 }
