+++ src/main/kotlin/com/admin_bot/features/bot/data/BotInfo.kt
@@ -1,8 +1,8 @@
 package com.admin_bot.features.bot.data
 
 data class BotInfo(
-    val id: Int,
+    val id: Long,
     val token: String,
     val adminEmail: String? = null,
-    val chatIds: Set<Int> = mutableSetOf(),
+    val chatIds: Set<Long> = mutableSetOf(),
 )
