+++ src/main/kotlin/com/admin_bot/features/bot/repository/BotRepository.kt
@@ -5,12 +5,12 @@ import com.admin_bot.features.bot.data.BotInfo
 import com.admin_bot.features.classification.data.MessageCategory
 
 interface BotRepository {
-    suspend fun getBotInfo(botId: Int): BotInfo
-    suspend fun addAdminEmail(botId: Int, email: String)
-    suspend fun addChat(botId: Int, chatId: Int)
-    suspend fun deleteChat(botId: Int, chatId: Int)
+    suspend fun getBotInfo(botId: Long): BotInfo
+    suspend fun addAdminEmail(botId: Long, email: String)
+    suspend fun addChat(botId: Long, chatId: Long)
+    suspend fun deleteChat(botId: Long, chatId: Long)
 
-    suspend fun setAction(messagesCategory: MessageCategory, botAction: BotAction, botId: Int)
+    suspend fun setAction(messagesCategory: MessageCategory, botAction: BotAction, botId: Long)
 
     suspend fun clear()
 }
