+++ src/main/kotlin/com/admin_bot/features/chat/data/ChatInfo.kt
@@ -1,6 +1,6 @@
 package com.admin_bot.features.chat.data
 
-data class ChatInfo(val id: Int, val title: String, val type: ChatType)
+data class ChatInfo(val id: Long, val title: String, val type: ChatType)
 
 enum class ChatType {
     GROUP,
