+++ src/main/kotlin/com/admin_bot/features/chat/repository/ChatRepository.kt
@@ -5,11 +5,11 @@ import com.admin_bot.features.chat.data.ChatInfo
 import com.admin_bot.features.chat.data.ChatType
 
 interface ChatRepository {
-    suspend fun getChatInfo(chatId: Int): ChatInfo
+    suspend fun getChatInfo(chatId: Long): ChatInfo
     suspend fun getAllChatsInfo(): List<ChatInfo>
-    suspend fun getAllChatsInfo(botId: Int): List<ChatInfo>
+    suspend fun getAllChatsInfo(botId: Long): List<ChatInfo>
 
     suspend fun addChat(chatDescription: ChatDescription)
-    suspend fun deleteChatInfo(chatId: Int)
+    suspend fun deleteChatInfo(chatId: Long)
     suspend fun clear()
 }
