+++ src/main/kotlin/com/admin_bot/features/classification/repository/ClassificationHistoryRepository.kt
@@ -3,6 +3,6 @@ package com.admin_bot.features.classification.repository
 import com.admin_bot.features.classification.data.MessageCategory
 
 interface ClassificationHistoryRepository {
-    suspend fun saveAdminDecision(messageId: Int, category: MessageCategory)
-    suspend fun saveBotDecision(messageId: Int, category: MessageCategory)
+    suspend fun saveAdminDecision(messageId: Long, category: MessageCategory)
+    suspend fun saveBotDecision(messageId: Long, category: MessageCategory)
 }
