+++ src/main/kotlin/com/admin_bot/features/messages/data/GetMessagesParams.kt
@@ -2,9 +2,9 @@ package com.admin_bot.features.messages.data
 
 import io.ktor.http.*
 
-data class GetMessagesParams(val pageNumber: Int?, val pageSize: Int?) {
+data class GetMessagesParams(val pageNumber: Long?, val pageSize: Long?) {
     constructor(parameters: Parameters) : this(
-        pageNumber = parameters["pageNumber"]?.toInt(),
-        pageSize = parameters["pageSize"]?.toInt()
+        pageNumber = parameters["pageNumber"]?.toLong(),
+        pageSize = parameters["pageSize"]?.toLong()
     )
 }
