+++ src/main/kotlin/com/admin_bot/features/messages/repository/SavedMessagesRepository.kt
@@ -3,9 +3,9 @@ package com.admin_bot.features.messages.repository
 import com.admin_bot.features.messages.data.Message
 
 interface SavedMessagesRepository {
-    suspend fun getMessages(botId: Int, pageNumber: Int?, pageSize: Int?)
-    suspend fun getMessages(botId: Int, chatId: Int, pageNumber: Int?, pageSize: Int?)
-    suspend fun addMessage(botId: Int, message: Message)
-    suspend fun deleteMessage(messageId: Int)
+    suspend fun getMessages(botId: Long, pageNumber: Long?, pageSize: Long?)
+    suspend fun getMessages(botId: Long, chatId: Long, pageNumber: Long?, pageSize: Long?)
+    suspend fun addMessage(botId: Long, message: Message)
+    suspend fun deleteMessage(messageId: Long)
     suspend fun clear()
 }
