+++ src/main/kotlin/com/admin_bot/features/user/data/UserInfo.kt
@@ -1,3 +1,3 @@
 package com.admin_bot.features.user.data
 
-data class UserInfo(val id: Int, val firstName: String, val lastName: String)
+data class UserInfo(val id: Long, val firstName: String, val lastName: String)
