+++ src/main/kotlin/com/admin_bot/features/user/repository/UserRepository.kt
@@ -5,11 +5,11 @@ import com.admin_bot.features.user.data.UserInfo
 
 
 interface UserRepository {
-    suspend fun getUserInfo(UserId: Int): UserInfo
+    suspend fun getUserInfo(UserId: Long): UserInfo
     suspend fun getAllUsersInfo(): List<UserInfo>
-    suspend fun getAllUsersInfo(botId: Int): List<UserInfo>
+    suspend fun getAllUsersInfo(botId: Long): List<UserInfo>
 
     suspend fun addUser(userDescription: UserDescription)
-    suspend fun deleteUserInfo(UserId: Int)
+    suspend fun deleteUserInfo(UserId: Long)
     suspend fun clear()
 }
