+++ src/main/kotlin/com/admin_bot/plugins/mocks/model/authentication/MockRefreshTokenStorage.kt
@@ -3,9 +3,9 @@ package com.admin_bot.plugins.mocks.model.authentication
 import com.admin_bot.features.authentification.model.RefreshTokenStorage
 
 class MockRefreshTokenStorage: RefreshTokenStorage {
-    private val refreshTokens: MutableMap<Int, MutableSet<String>> = mutableMapOf()
+    private val refreshTokens: MutableMap<Long, MutableSet<String>> = mutableMapOf()
 
-    override suspend fun saveRefreshToken(refreshToken: String, botId: Int) {
+    override suspend fun saveRefreshToken(refreshToken: String, botId: Long) {
         if (refreshTokens.containsKey(botId)) {
             refreshTokens[botId]!!.add(refreshToken)
         } else {
@@ -13,7 +13,7 @@ class MockRefreshTokenStorage: RefreshTokenStorage {
         }
     }
 
-    override suspend fun deleteRefreshTokenIfContains(refreshToken: String, botId: Int): Boolean {
+    override suspend fun deleteRefreshTokenIfContains(refreshToken: String, botId: Long): Boolean {
         val refreshTokensSet = refreshTokens[botId]
         if (refreshTokensSet?.contains(refreshToken) == true) {
             refreshTokensSet.remove(refreshToken)
